Refactor difm purchase details function
